Complete Client CRUD with DTO and service exception handling

+++ src/main/java/com/josev001/dsclient/services/ClientService.java
@@ -3,10 +3,14 @@ package com.josev001.dsclient.services;
 import com.josev001.dsclient.dto.ClientDTO;
 import com.josev001.dsclient.entities.Client;
 import com.josev001.dsclient.repositories.ClientRepository;
+import com.josev001.dsclient.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ClientService {
@@ -15,9 +19,9 @@ public class ClientService {
     private ClientRepository repository;
 
     public ClientDTO findById(Long id) {
-        Client entity = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Cliente não encontrado!"));
-        return new ClientDTO(entity);
+            Client entity = repository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado"));
+            return new ClientDTO(entity);
     }
 
     public Page<ClientDTO> findAll(Pageable pageable) {
@@ -26,21 +30,29 @@ public class ClientService {
     }
 
     public ClientDTO insert(ClientDTO dto) {
-        Client entity = new Client();
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+            Client entity = new Client();
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
     }
 
     public ClientDTO update(Long id, ClientDTO dto) {
+        try {
         Client entity = repository.getReferenceById(id);
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
         return new ClientDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Cliente não encontrado");
+        }
     }
 
+
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente não encontrado");
+        }
+            repository.deleteById(id);
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
